Переносит всю логику по отрисовке элементов доски задач в контроллер board-controller
